fix(pr-execute): address 4 findings from triage
- phase8-stale-case-authored-comment: the step-4 comment in
  case_proposal_received_tree.py claimed ("Add","CaseStatus") was NOT in
  _CASE_AUTHORED_SIGNATURES; this PR added it per CLP-12-001. Reworded to
  say the vendor URI is a provenance choice, not a guard constraint.
- phase8-stale-assert-message: same staleness in the
  test_add_case_status_uses_vendor_actor assert message.
- phase9-stale-chain-py-docstring: test_close_case_routing.py pointed at
  nodes/chain.py for the _CASE_AUTHORED_SIGNATURES entry; the table moved
  to nodes/canonical_entry.py in this PR.
- phase10-canonical-entry-tests-not-mirrored: added
  test/core/behaviors/sync/nodes/test_canonical_entry.py mirroring the new
  module (BTND-07-002, CS-18-004) and moved the 7 direct-validator tests
  out of test_chain.py. Node-level tests that exercise the guard through
  CreateLogEntryNode stay in test_chain.py.
